Added 3 new files...
rgb-class.js

creates the class for the rgb party effect object
contains the code for setting the interval and
the data for the array, element id, etc

rgb-usage.js

this simulates a file that is used after you import
the rgb-class.js.  the idea is that a progammer
can import rgb-class.js and then in his code he
can simply create rgb classes.

test.html

the page i'll use to open with my browser to test
future javascript widgets